+++ Functionality/Package.cs
@@ -8,7 +8,7 @@ namespace MailLogic
 {
     public class Package : Mail
     {
-        public Package(int weight, int id)
+        public Package(double weight, int id)
         {
             this.weight = 0;
             this.id = 0;
